+++ src/subcommands/mod.rs
@@ -1,9 +1,8 @@
 pub mod add;
 pub mod list;
+pub mod modpack;
 pub mod profile;
 mod remove;
-mod switch;
 mod upgrade;
 pub use remove::remove;
-pub use switch::switch;
 pub use upgrade::upgrade;