+++ src/subcommands/upgrade.rs
@@ -1,14 +1,13 @@
-use crate::{mutex_ext::MutexExt, CROSS, STYLE, TICK, YELLOW_TICK};
-use anyhow::{bail, Error, Result};
+use crate::{download::download, CROSS, STYLE, TICK, YELLOW_TICK};
+use anyhow::{bail, Result};
 use colored::Colorize;
 use ferinth::Ferinth;
 use fs_extra::file::{move_file, CopyOptions};
 use furse::Furse;
 use indicatif::ProgressBar;
 use itertools::Itertools;
-use libium::{check, config, upgrade};
+use libium::{config::structs::Profile, mutex_ext::MutexExt, upgrade::mod_downloadable};
 use octocrab::Octocrab;
-use size::{Base, Size, Style};
 use std::{
     fs::read_dir,
     sync::{
@@ -16,16 +15,13 @@ use std::{
         Arc, Mutex,
     },
 };
-use tokio::{
-    fs::{copy, remove_file},
-    spawn,
-};
+use tokio::{fs::remove_file, spawn};
 
 pub async fn upgrade(
     modrinth: Arc<Ferinth>,
     curseforge: Arc<Furse>,
     github: Arc<Octocrab>,
-    profile: &config::structs::Profile,
+    profile: &Profile,
 ) -> Result<()> {
     let profile = Arc::new(profile.clone());
     let to_download = Arc::new(Mutex::new(Vec::new()));
@@ -51,7 +47,7 @@ pub async fn upgrade(
         let github = github.clone();
         let mod_ = mod_.clone();
         tasks.push(spawn(async move {
-            let result = upgrade::get_latest_compatible_downloadable(
+            let result = mod_downloadable::get_latest_compatible_downloadable(
                 modrinth.clone(),
                 curseforge.clone(),
                 github.clone(),
@@ -80,7 +76,9 @@ pub async fn upgrade(
                     }
                 },
                 Err(err) => {
-                    if let upgrade::Error::ModrinthError(ferinth::Error::RateLimitExceeded(_)) = err
+                    if let mod_downloadable::Error::ModrinthError(
+                        ferinth::Error::RateLimitExceeded(_),
+                    ) = err
                     {
                         // Immediately fail if there is a rate limit
                         progress_bar.finish_and_clear();
@@ -151,56 +149,12 @@ pub async fn upgrade(
         }
     }
 
-    if to_download.is_empty() && to_install.is_empty() {
-        println!("\n{}", "All up to date!".bold());
-    } else {
-        println!("\n{}\n", "Downloading Mod Files".bold());
-        let progress_bar = Arc::new(Mutex::new(
-            ProgressBar::new(to_download.len() as u64).with_style(STYLE.clone()),
-        ));
-        {
-            progress_bar.force_lock().enable_steady_tick(100);
-        }
-        let mut tasks = Vec::new();
-        for downloadable in to_download {
-            let progress_bar = progress_bar.clone();
-            let downloadable = downloadable.clone();
-            let profile = profile.clone();
-            tasks.push(spawn(async move {
-                let contents = reqwest::get(&downloadable.download_url)
-                    .await?
-                    .bytes()
-                    .await?;
-                let size = Size::Bytes(contents.len());
-                check::write_mod_file(&profile.output_dir, contents, &downloadable.filename)
-                    .await?;
-                let progress_bar = progress_bar.force_lock();
-                progress_bar.println(format!(
-                    "{} Downloaded {:7} {}",
-                    &*TICK,
-                    size.to_string(Base::Base10, Style::Smart),
-                    downloadable.filename.dimmed(),
-                ));
-                progress_bar.set_position(progress_bar.position() + 1);
-                Ok::<(), Error>(())
-            }));
-        }
-        for handle in tasks {
-            handle.await??;
-        }
-        Arc::try_unwrap(progress_bar)
-            .expect("Failed to run threads to completion")
-            .into_inner()?
-            .finish_and_clear();
-        for installable in to_install {
-            eprint!(
-                "Installing  {}... ",
-                installable.0.to_string_lossy().dimmed()
-            );
-            copy(installable.1, profile.output_dir.join(installable.0)).await?;
-            println!("{}", &*TICK);
-        }
-    }
+    download(
+        Arc::new(profile.output_dir.clone()),
+        to_download,
+        to_install,
+    )
+    .await?;
 
     if error.load(Ordering::Relaxed) {
         bail!("\nCould not get the latest compatible version of some mods")